Add header field.
One use of the log structure is to be passed between threads, from
a generating thread to the actual logging thread.  It's likely that
in such a situation, there'll be other structures being passed via
the same path (eg. queue), so change the first four bytes to be a
field that can be used to identify log "messages".

Clean up some doxygen; make comments more consistent; etc.