Improve the debian packages.
We now have a runtime and independent -dev package for
ubuntu/debian. We share a little more comonality at the Makefile level
with the rpm build. We've moved to compat level 9 although this has
semingly caused a need to enumerate the files vis dh_movefiles now we
have two packages.

We also now build in a subdirectory which means that whilst we
configure an extra time we don't break the configured env for the
current build.